% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fit.R
\name{heatmap_fit}
\alias{heatmap_fit}
\title{Heatmap fit function}
\usage{
heatmap_fit(property, eset, precomp_dir, datasetname, clustering_methods)
}
\arguments{
\item{property}{Property or the variable of interest}

\item{eset}{}

\item{precomp_dir}{}

\item{datasetname}{}

\item{clustering_methods}{Which method to use for clustering.}
}
\description{
This function allows you to draw heatmap that fits to the screen using ComplexHeatmap package.
}
\examples{
heatmap_fit(property="ER",eset, precomp_dir,datasetname="EDS-1013", clustering_methods="Clustering by groups")
}
