% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_scroll.R
\name{heatmap_scroll}
\alias{heatmap_scroll}
\title{Heatmap scroll function}
\usage{
heatmap_scroll(property, eset, precomp_dir, datasetname, clustering_methods)
}
\arguments{
\item{property}{Property or the variable of interest}

\item{eset}{}

\item{precomp_dir}{}

\item{datasetname}{}

\item{clustering_methods}{Which method to use for clustering.}
}
\description{
This function allows you to draw heatmap that let you scroll down with the top most highly variable genes using ComplexHeatmap package.
}
\examples{
heatmap_fit(property="ER",eset, precomp_dir,datasetname="EDS-1013", clustering_methods="Clustering by groups")
}
