% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2D.R
\name{PCA2D}
\alias{PCA2D}
\title{PCA2D function}
\usage{
PCA2D(eset, logTransform = FALSE, property)
}
\arguments{
\item{eset}{}

\item{logTransform}{If the data should be log transformed? Default is FALSE.}

\item{property}{Property or the variable of interest}
}
\description{
This function allows you to draw 2D PCA plot.
}
\examples{
PCA2D(eset,logTransform=FALSE, property="ER")
}
